Separated scraper methods from Ambassador class, tests still pass
